chore(routing): moving event mod into routing_api mod
