+++ src/routing/routing_api/event_stream.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::routing::event::Event;
+use super::Event;
 use std::fmt::{self, Debug, Formatter};
 use tokio::sync::mpsc;
 /// Stream of routing node events
